not include buffer into browser bundle
