+++ index.js
@@ -1,3 +1,3 @@
 'use strict';
 
-module.exports = require('./src');
+module.exports = require('./src/node');
