--- src/index.js
@@ -1,14 +1,0 @@
-'use strict';
-
-const isICO = require('./is-ico');
-const parse = require('./parse');
-
-/**
- * @class ICO
- */
-const ICO = {
-  isICO,
-  parse
-};
-
-module.exports = ICO;
