+++ src/node/buffer-to-arraybuffer.js
