+++ src/parse-ico.js
@@ -1,6 +1,5 @@
 'use strict';
 
-const bufferToArrayBuffer = require('./utils/buffer-to-arraybuffer');
 const bitDepthOfPNG = require('./bit-depth-of-png');
 const isCUR = require('./is-cur');
 const isICO = require('./is-ico');
@@ -11,7 +10,7 @@ const range = require('./utils/range');
 /**
  * Parse ICO and return some image object.
  * @access private
- * @param {ArrayBuffer|Buffer} buffer ICO file data.
+ * @param {ArrayBuffer} arrayBuffer ICO file data.
  * @param {String} mime MIME type for output.
  * @param {Object} Image Image encoder/decoder
  * @returns {Promise<Object[]>} Resolves to array of parsed ICO.
@@ -20,10 +19,9 @@ const range = require('./utils/range');
  *   * `bit` **Number** - Image bit depth.
  *   * `buffer` **ArrayBuffer** - Image buffer.
  */
-const parseICO = (buffer, mime, Image) => {
-  const arrayBuffer = bufferToArrayBuffer(buffer);
-  if (!arrayBuffer) {
-    return Promise.reject(new TypeError('"buffer" argument must be a Buffer or ArrayBuffer'));
+const parseICO = (arrayBuffer, mime, Image) => {
+  if (!(arrayBuffer instanceof ArrayBuffer)) {
+    return Promise.reject(new TypeError('"buffer" argument must be an ArrayBuffer'));
   }
   if (!isCUR(arrayBuffer) && !isICO(arrayBuffer)) {
     return Promise.reject(new Error('buffer is not ico'));
