+++ test/is-png.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const bufferToArrayBuffer = require('../src/utils/buffer-to-arraybuffer');
+const bufferToArrayBuffer = require('../src/node/buffer-to-arraybuffer');
 const chai = require('chai');
 const fs = require('fs');
 const isPNG = require('../src/is-png');
